+++ src/Rocks/Builders/Create/MockEventExtensionsBuilder.cs
@@ -7,7 +7,8 @@ namespace Rocks.Builders.Create;
 
 internal static class MockEventExtensionsBuilder
 {
-	internal static void Build(IndentedTextWriter writer, TypeMockModel mockType, ImmutableHashSet<string> adornmentsFQNs)
+	internal static void Build(IndentedTextWriter writer, TypeMockModel mockType, 
+		ImmutableHashSet<(string adornmentFQN, string typeArguments, string constraints)> adornmentsFQNs)
 	{
 		writer.WriteLines(
 			$$"""
@@ -19,11 +20,11 @@ internal static class MockEventExtensionsBuilder
 
 		foreach (var (name, argsType) in mockType.Events.Select(_ => (_.Name, _.ArgsType)).Distinct())
 		{
-			foreach (var adornment in adornmentsFQNs.OrderBy(_ => _))
+			foreach ((var adornmentFQN, var typeArguments, var constraints) in adornmentsFQNs.OrderBy(_ => _))
 			{
 				writer.WriteLines(
 					$$"""
-					internal static {{adornment}} Raise{{name}}(this {{adornment}} self, {{argsType}} args) => 
+					internal static {{adornmentFQN}} Raise{{name}}{{typeArguments}}(this {{adornmentFQN}} self, {{argsType}} args){{constraints}} => 
 						self.AddRaiseEvent(new("{{name}}", args));
 					""");
 			}