+++ src/Rocks/Builders/Create/PropertyExpectationsPropertyBuilder.cs
@@ -9,7 +9,7 @@ namespace Rocks.Builders.Create;
 internal static class PropertyExpectationsPropertyBuilder
 {
 	private static void BuildGetter(IndentedTextWriter writer, PropertyModel property, uint memberIdentifier, string expectationsFullyQualifiedName, 
-		Action<string> adornmentsFQNsPipeline)
+		Action<string, string, string> adornmentsFQNsPipeline)
 	{
 		var propertyGetMethod = property.GetMethod!;
 		var callbackDelegateTypeName = propertyGetMethod.RequiresProjectedDelegate ?
@@ -35,7 +35,7 @@ internal static class PropertyExpectationsPropertyBuilder
 			adornmentsType = $"global::Rocks.Adornments<{handlerTypeName}, {callbackDelegateTypeName}, {returnType}>";
 		}
 
-		adornmentsFQNsPipeline(adornmentsType);
+		adornmentsFQNsPipeline(adornmentsType, string.Empty, string.Empty);
 
 		writer.WriteLines(
 			$$"""
@@ -51,7 +51,7 @@ internal static class PropertyExpectationsPropertyBuilder
 	}
 
 	private static void BuildSetter(IndentedTextWriter writer, PropertyModel property, uint memberIdentifier, string expectationsFullyQualifiedName, 
-		Action<string> adornmentsFQNsPipeline)
+		Action<string, string, string> adornmentsFQNsPipeline)
 	{
 		var propertyParameterType = property.SetMethod!.Parameters[0].Type;
 		var propertyParameterValue =
@@ -64,7 +64,7 @@ internal static class PropertyExpectationsPropertyBuilder
 			MockProjectedDelegateBuilder.GetProjectedCallbackDelegateFullyQualifiedName(property.SetMethod!, property.MockType) :
 			DelegateBuilder.Build(property.SetMethod!.Parameters);
 		var adornmentsType = $"global::Rocks.Adornments<{expectationsFullyQualifiedName}.Handler{memberIdentifier}, {callbackDelegateTypeName}>";
-		adornmentsFQNsPipeline(adornmentsType);
+		adornmentsFQNsPipeline(adornmentsType, string.Empty, string.Empty);
 
 		writer.WriteLines(
 			$$"""
@@ -86,7 +86,7 @@ internal static class PropertyExpectationsPropertyBuilder
 	}
 
 	internal static void Build(IndentedTextWriter writer, PropertyModel property, PropertyAccessor accessor, string expectationsFullyQualifiedName,
-		Action<string> adornmentsFQNsPipeline)
+		Action<string, string, string> adornmentsFQNsPipeline)
 	{
 		var memberIdentifier = property.MemberIdentifier;